+++ CHANGELOG.md
@@ -25,8 +25,13 @@ Add a link to the GitHub diff like
 
 [Unreleased]: https://github.com/winbench/bench/compare/master...dev
 
+### Added
+* Config property `QuickAccessPowerShellCore`
+* PowerShell Core button in quick access toolbar
+
 ### Changed
 * New bash icon in quick access toolbar
+* Hotkey for Windows PowerShell changed from Alt+P to Alt+W
 
 ## [0.22.4] - 2020-01-06